Move TestSpanData to a testing package instead of in the published SDK.
