+++ exporters/inmemory/build.gradle
@@ -13,6 +13,8 @@ dependencies {
 
     annotationProcessor libraries.auto_value
 
+    testImplementation project(':opentelemetry-testing-internal')
+
     signature "org.codehaus.mojo.signature:java17:1.0@signature"
     signature "net.sf.androidscents.signature:android-api-level-24:7.0_r2@signature"
 }
