+++ exporters/logging/build.gradle
@@ -11,6 +11,8 @@ ext.moduleName = "io.opentelemetry.exporters.logging"
 dependencies {
     api project(':opentelemetry-sdk')
 
+    testImplementation project(':opentelemetry-testing-internal')
+    
     signature "org.codehaus.mojo.signature:java17:1.0@signature"
     signature "net.sf.androidscents.signature:android-api-level-24:7.0_r2@signature"
 }
