+++ exporters/otlp/build.gradle
@@ -18,6 +18,8 @@ dependencies {
             libraries.protobuf,
             libraries.protobuf_util
 
+    testImplementation project(':opentelemetry-testing-internal')
+
     testImplementation "io.grpc:grpc-testing:${grpcVersion}"
     testRuntime "io.grpc:grpc-netty-shaded:${grpcVersion}"
