+++ exporters/zipkin/build.gradle
@@ -18,6 +18,7 @@ dependencies {
     implementation libraries.zipkin_reporter,
             libraries.zipkin_okhttp
 
+    testImplementation project(':opentelemetry-testing-internal')
     testImplementation libraries.guava,
             libraries.zipkin_junit
