+++ sdk/build.gradle
@@ -20,7 +20,10 @@ dependencies {
     testAnnotationProcessor libraries.auto_value
     testCompileOnly libraries.auto_value_annotation
 
+    testImplementation project(':opentelemetry-testing-internal')
     testImplementation libraries.junit_pioneer
+
+    jmh project(':opentelemetry-testing-internal')
 
     signature "org.codehaus.mojo.signature:java17:1.0@signature"
     signature "net.sf.androidscents.signature:android-api-level-24:7.0_r2@signature"
