+++ testing_internal/src/main/java/io/opentelemetry/sdk/trace/TestSpanData.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019, OpenTelemetry Authors
+ * Copyright 2020, OpenTelemetry Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.sdk.trace.data.test;
+package io.opentelemetry.sdk.trace;
 
 import com.google.auto.value.AutoValue;
 import io.opentelemetry.common.AttributeValue;
